AB#66247 - Merge bug/66247 to develop

+++ src/ESFA.DC.ILR1819.DataStore.Database/Rulebase/Tables/TBL_LearningDelivery.sql
@@ -30,7 +30,7 @@
     [AchievementApplicVal]           DECIMAL (10, 5) NULL,
     [AchPayment]                     DECIMAL (10, 5) NULL,
     [ActualNumInstalm]               INT             NULL,
-    [CombinedAdjProp]                BIGINT          NULL,
+    [CombinedAdjProp]                DECIMAL (10, 5) NULL,
     [EmpIdAchDate]                   INT             NULL,
     [LearnDelDaysIL]                 INT             NULL,
     [LearnDelStandardAccDaysIL]      INT             NULL,
